+++ egov-workflow-v2/CHANGELOG.md
@@ -3,6 +3,16 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.2.0 2021-06-23
+- Added autoescalation feature
+- Added statelevel fallback feature at businessService level
+
+## 1.1.5 2021-05-11
+- Inbox and processInstance count for each status
+
+## 1.1.4 2021-03-17
+- Bug fix for rate action validation
+
 ## 1.1.3 2021-02-26
 - Updated domain name in application.properties
 - Added rating field in processInstance